refactor(crypto): move inline merkle-backend tests to tests/
Relocate #[cfg(test)] mod tests blocks out of crypto/crypto source
files into crypto/crypto/src/tests/.

- backends/field_element.rs tests -> tests/field_element_tests.rs
- backends/field_element_vector.rs tests -> tests/field_element_vector_tests.rs
- merkle.rs disk_spill_serde_tests -> appended to tests/merkle_tests.rs
- MerkleTree::mmap_backing bumped to pub(crate) so the relocated
  disk-spill serde test can reach it
