+++ crypto/crypto/src/tests/mod.rs
@@ -1,4 +1,6 @@
 pub mod default_transcript_tests;
+pub mod field_element_tests;
+pub mod field_element_vector_tests;
 pub mod merkle_proof_tests;
 pub mod merkle_tests;
 pub mod merkle_utils_tests;
